Adding in a __debug__.require function when debugging react-native
In order to enable projects with native components to provide fallbacks using node modules, this change exposes __debug__.require which acts as if a file in /.vscode/.react/ invokes require(). This means that it can access any native module (such as fs or child_process) or any installed package in a parent folder.